separate tests depending on precondition
